+++ src/main/java/me/pajic/tiered_backpacks/util/BackpackClientUtil.java
@@ -11,12 +11,17 @@ public class BackpackClientUtil {
 
 	public static void onClientTick(Minecraft client) {
 		if (ModKeybinds.OPEN_BACKPACK.consumeClick() && client.player != null && client.level != null) {
-			ItemStack chestItem = client.player.getItemBySlot(EquipmentSlot.CHEST);
-			if (BackpackUtil.isValidContainerHolder(chestItem)) {
-				TieredBackpacks.xplat().sendToServer(new ModNetworking.C2SOpenBackpackPayload(1));
-			} else if (TieredBackpacks.CONFIG.canOpenFromInventory.get()) {
-				if (client.player.getInventory().getNonEquipmentItems().stream().anyMatch(stack -> stack.is(BackpackUtil.BACKPACKS))) {
-					TieredBackpacks.xplat().sendToServer(new ModNetworking.C2SOpenBackpackPayload(0));
+			ItemStack chestItem = ItemStack.EMPTY;
+			if (CompatFlags.TRINKETS_LOADED) chestItem = TrinketsCompat.tryGetTrinketBackpack(client.player);
+			if (!chestItem.isEmpty()) TieredBackpacks.xplat().sendToServer(new ModNetworking.C2SOpenBackpackPayload(2));
+			else {
+				chestItem = client.player.getItemBySlot(EquipmentSlot.CHEST);
+				if (BackpackUtil.isValidContainerHolder(chestItem)) {
+					TieredBackpacks.xplat().sendToServer(new ModNetworking.C2SOpenBackpackPayload(1));
+				} else if (TieredBackpacks.CONFIG.canOpenFromInventory.get()) {
+					if (client.player.getInventory().getNonEquipmentItems().stream().anyMatch(stack -> stack.is(BackpackUtil.BACKPACKS))) {
+						TieredBackpacks.xplat().sendToServer(new ModNetworking.C2SOpenBackpackPayload(0));
+					}
 				}
 			}
 		}