+++ src/main/java/me/pajic/tiered_backpacks/util/TrinketsCompat.java
@@ -1,0 +1,14 @@
+package me.pajic.tiered_backpacks.util;
+
+import eu.pb4.trinkets.api.TrinketsApi;
+import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.item.ItemStack;
+
+public class TrinketsCompat {
+
+	public static ItemStack tryGetTrinketBackpack(LivingEntity entity) {
+		var list = TrinketsApi.getAttachment(entity).getEquipped(stack -> stack.is(BackpackUtil.BACKPACKS));
+		if (!list.isEmpty()) return list.getFirst().getB();
+		return ItemStack.EMPTY;
+	}
+}